requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-leaflet>=0.29
django-bootstrap5>=23.3
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.5
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
plotly>=5.18
joblib>=1.3
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0
